pyproject.toml
[tool.poetry]
name = "telegram-fashion-bot"
version = "0.1.0"
description = ""
authors = ["mannyuncharted <omoebun52@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.12"
python-telegram-bot = {extras = ["job-queue"], version = "^21.9"}
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.12"
langchain-google-genai = "^2.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
